fixed compiler warnings under visual studio (2)

+++ source/converter/inc/SampleBuffer.hpp
@@ -91,7 +91,7 @@ void SampleBuffer<sample_base_t>::DoAddSample( sample_base_t x )
    if( mBufferPos >= static_cast<int32_t>(mSampleBuffer.size()) )
    {
       mSampleBuffer.resize( mSampleBuffer.size() + BASE_BUFFER_SIZE_IN_BYTES );
-      mBufferSize = mSampleBuffer.size();
+      mBufferSize = static_cast<int32_t>( mSampleBuffer.size());
       
    }
